requirements.txt
# Core vector store


# Numeric
numpy>=1.25.0,<1.26.0

# PyTorch CPU builds
torch==2.2.2
torchvision==0.17.2

# Transformers + Sentence-Transformers
transformers>=4.41.0,<5.0.0
sentence-transformers==2.2.2

# Hugging Face Hub (restores cached_download)
huggingface_hub==0.25.2

# Tokenizers & NLP utilities
sentencepiece
nltk



langchain-community

faiss-cpu==1.7.4
langchain==0.2.0
PyPDF2==3.0.1

# Streamlit for UI
streamlit==1.35.0

# Utilities

tqdm>=4.66.4

# Optional: if you use HuggingFace LLaMA3 or similar (for LLaMA3 swap-in later)
accelerate>=0.30.0
openai==0.28
python-dotenv
cerebras_cloud_sdk